Lab 4 - GCP Labs/requirements.txt
flask>=2.0.0
gunicorn>=21.0.0
pandas>=1.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
joblib>=1.0.0
google-cloud-storage>=2.0.0
google-cloud-bigquery>=3.0.0
